tests/wpt/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.create1.zero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.create1.zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.imageData.create1.zero</h1>
<p class="desc">createImageData(null) throws TypeError</p>


<script>
var t = async_test("createImageData(null) throws TypeError");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

assert_throws(new TypeError(), function() { ctx.createImageData(null); });
t.done();

});
</script>
